// frontend/src/modules/cinema/models/movie.ts
export interface IMovieItem {
    movieId: number 
    movieName: string 
    duration: number
    imageURL: string 
    movieGenre: string 
}

export interface IMovie {
    movieId: number 
    movieName: string 
    duration: number
    imageURL: string 
    description: string 
    releaseDate: string 
    movieGenre: string 
}